Fix scenario 1
